requirements.txt
langchain-ollama>=0.1.0
langchain-google-genai>=0.0.1
python-dotenv>=1.0.0
mcp==1.26.0
mcp-use
aiohttp
geopy
testcontainers